Complete q4 iterative and recursiver version
I implement the Kadane's algorithm in iterative version to solve how to find the max subarray.
I use the concept of the divide and conquer in recursive version,this algorithm will break down a array into
two or more subarray of the same type,and I compare the left value,right value,and max sum of subarray from mid
to left and mid to right,so I can get the max value from these 3 condition.
When the end of the program execution,I will show the sum of the max subbarray,and the composition of the subarray
in both methods.