Complete q5 iterative and recursive version
I use the a ListNode type array(named record) to become the traversal table,and make the *ptr->next to check the
traversal table in each recursive call,until find the address from traversal table,then it means that I find the cycle begin,so I return the address,if not, then return NULL,this method implemented in the recursive version.
I implement the Floyd's cycle-finding algorithm(also called the "tortoise and the hare algorithm") in the
iterativer version,the tortoise speed is 1x and the hare speed is 2x,they will start run from the same point
and if there is a cycle in this linklist, then the tortoise and hare will meet at point inside the cycle,when I
find the meet point,I make them run again from the different point(one start from the Linklist head,the other start
from the meet point),and they will meet each other again at the point inside the cycle,and this point will be
the cycle begin point.